Add files via upload
We see that we are no more successful and no more able to access the admin account. The error indicates that there was no user with credentials username admin'# and password admin.
A prepared statement goes through the compilation step and turn into a pre compiled query with empty placeholder for data.